Add base HistoryActivity for testing
