+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,7 @@
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
+                <action android:name="android.intent.action.VIEW"/>
             </intent-filter>
         </activity>
 
@@ -109,6 +110,8 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".MainActivity"/>
         </activity>
+
+        <activity android:name=".activity.HistoryActivity"/>
 
         <!-- Testing -->
         <activity android:name=".test.TestNavigationActivity"/>
