+++ app/src/main/java/com/mapbox/services/android/navigation/testapp/activity/navigationui/ComponentNavigationActivity.java
@@ -14,7 +14,6 @@ import android.support.design.widget.BaseTransientBottomBar;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.transition.TransitionManager;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
 import com.mapbox.android.core.location.LocationEngine;
@@ -32,6 +31,7 @@ import com.mapbox.mapboxsdk.maps.MapView;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback;
 import com.mapbox.services.android.navigation.testapp.R;
+import com.mapbox.services.android.navigation.testapp.activity.HistoryActivity;
 import com.mapbox.services.android.navigation.testapp.activity.location.FusedLocationEngine;
 import com.mapbox.services.android.navigation.ui.v5.camera.DynamicCamera;
 import com.mapbox.services.android.navigation.ui.v5.instruction.InstructionView;
@@ -48,8 +48,6 @@ import com.mapbox.services.android.navigation.v5.offroute.OffRouteListener;
 import com.mapbox.services.android.navigation.v5.routeprogress.ProgressChangeListener;
 import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.List;
 import java.util.Locale;
 
@@ -61,7 +59,7 @@ import retrofit2.Callback;
 import retrofit2.Response;
 import timber.log.Timber;
 
-public class ComponentNavigationActivity extends AppCompatActivity implements OnMapReadyCallback,
+public class ComponentNavigationActivity extends HistoryActivity implements OnMapReadyCallback,
   MapboxMap.OnMapLongClickListener, LocationEngineListener, ProgressChangeListener,
   MilestoneEventListener, OffRouteListener {
 
@@ -77,8 +75,6 @@ public class ComponentNavigationActivity extends AppCompatActivity implements On
   private static final double DEFAULT_TILT = 0d;
   private static final double DEFAULT_BEARING = 0d;
   private static final int ONE_SECOND_INTERVAL = 1000;
-  private static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat("yyyy_MM_dd_HH_mm_ss");
-  private static final String JSON_EXTENSION = ".json";
 
   @BindView(R.id.componentNavigationLayout)
   ConstraintLayout navigationLayout;
@@ -103,7 +99,6 @@ public class ComponentNavigationActivity extends AppCompatActivity implements On
   private DirectionsRoute route;
   private Point destination;
   private MapState mapState;
-  private String filename;
 
   private enum MapState {
     INFO,
@@ -174,8 +169,6 @@ public class ComponentNavigationActivity extends AppCompatActivity implements On
 
     // Start navigation
     adjustMapPaddingForNavigation();
-    navigation.toggleHistory(true);
-    filename = buildFileName();
     navigation.startNavigation(route);
 
     // Location updates will be received from ProgressChangeListener
@@ -232,8 +225,6 @@ public class ComponentNavigationActivity extends AppCompatActivity implements On
 
   @Override
   public void onProgressChange(Location location, RouteProgress routeProgress) {
-    new StoreHistoryTask(navigation, filename).execute();
-
     // Cache "snapped" Locations for re-route Directions API requests
     updateLocation(location);
 
@@ -343,6 +334,7 @@ public class ComponentNavigationActivity extends AppCompatActivity implements On
     navigation.addMilestoneEventListener(this);
     navigation.addOffRouteListener(this);
     navigationMap.addProgressChangeListener(navigation);
+    addNavigationForHistory(navigation);
   }
 
   private void showSnackbar(String text, int duration) {
@@ -417,7 +409,6 @@ public class ComponentNavigationActivity extends AppCompatActivity implements On
     navigationMap.removeRoute();
     navigationMap.clearMarkers();
     navigation.stopNavigation();
-    navigation.toggleHistory(false);
     moveCameraOverhead();
   }
 
@@ -475,18 +466,5 @@ public class ComponentNavigationActivity extends AppCompatActivity implements On
     } else {
       vibrator.vibrate(ONE_HUNDRED_MILLISECONDS);
     }
-  }
-
-  private String buildFileName() {
-    StringBuilder sb = new StringBuilder();
-    sb.append(obtainCurrentTimeStamp());
-    sb.append(JSON_EXTENSION);
-    return sb.toString();
-  }
-
-  private String obtainCurrentTimeStamp() {
-    Date now = new Date();
-    String strDate = DATE_FORMAT.format(now);
-    return strDate;
   }
 }
